backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.5",
    "django-cors-headers>=4.6.0",
    "django-ninja>=1.3.0",
    "django-unfold>=0.46.0",
    "faker>=35.2.0",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "pytz>=2025.1",
    "rich>=13.9.4",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-django>=4.9.0",
    "ruff>=0.9.4",
]
